Move to separate demo.rs file, and make interactive.
